Bluetooth: Update time of response for adverisements
This patch updates the time stamp for the advertisment record
of a specfic bda, each time and advertisment is received.
This will help removal of the oldest record when the inquiry
database (of limt 40) is full. Without this patch, the 0th
element is removed whenever the the inq database gets filled
irrespective of whether that bda is the oldest updated or not.

CRs-Fixed: 972570
Change-Id: I1a12159de4839a76da1469fec8e13cccf97eb391