Clear notification registrations based on BD address
Notification registrations are cleared currently for all
the connected remote LE devices when anyone of the LE
devices are disconnected. Since the registrations are
cleared, the notifications from other LE devices which are
still connected are not processed by BT stack.
This patch clears notification registrations only for the
disconnected device based on Bluetooth device address
during GATT close.

CRs-Fixed: 2038482

Change-Id: Iae04a0266ba4efbf72f1527f7c2a057faafe9564